Resolve the PR against current LiteX while keeping the MIPS BIOS virtual-address mapping explicit and compatible with existing exports.

The merge keeps current CSR ordering/origin handling, adds VA aliases to generated memory/CSR headers through the CPU bios_map hook, and updates BIOS load paths to use VA addresses while preserving current load-size checks.

Also modernize the MIPS toolchain install hook, force little-endian code generation for GS232, and add export regression coverage for CPU BIOS address maps.

+++ litex/build/xilinx/yosys_nextpnr.py
@@ -19,8 +19,9 @@ from migen.fhdl.structure import _Fragment, wrap, Constant
 from migen.fhdl.specials import Instance
 
 from litex.build.yosys_nextpnr_toolchain import YosysNextPNRToolchain
+from litex.build.yosys_nextpnr_toolchain import yosys_nextpnr_args, yosys_nextpnr_argdict
 from litex.build.generic_platform import *
-from litex.build.xilinx.vivado import _xdc_separator, _format_xdc, _build_xdc
+from litex.build.xilinx.vivado import _xdc_separator, _format_xdc, _build_xdc, signed_bitstream_script
 from litex.build import tools
 from litex.build.xilinx import common
 
@@ -32,7 +33,9 @@ def _unwrap(value):
 # YosysNextpnrToolchain ----------------------------------------------------------------------------
 
 class XilinxYosysNextpnrToolchain(YosysNextPNRToolchain):
-    attr_translate = {}
+    attr_translate = {
+        "keep": ("keep", "true"),
+    }
 
     family     = "xilinx"
     synth_fmt  = "json"
@@ -68,6 +71,8 @@ class XilinxYosysNextpnrToolchain(YosysNextPNRToolchain):
     def _check_properties(self):
         pattern = re.compile("xc7([aksz])([0-9]+)(.*)-([0-9])")
         g = pattern.search(self.platform.device)
+        if g is None:
+            raise OSError(f"Unsupported device {self.platform.device}")
         if not self.dbpart:
             self.dbpart = f"xc7{g.group(1)}{g.group(2)}{g.group(3)}"
 
@@ -76,17 +81,27 @@ class XilinxYosysNextpnrToolchain(YosysNextPNRToolchain):
             self._xc7family = self.xc7_family_map[fam]
 
     def build_timing_constraints(self, vns):
-        xdc = []
+        max_freq = 0
+        xdc      = []
         xdc.append(_xdc_separator("Clock constraints"))
 
         for clk, [period, name] in sorted(self.clocks.items(), key=lambda x: x[0].duid):
             clk_sig = self._vns.get_name(clk)
+            # Search for the highest frequency.
+            freq = 1e3 / period
+            if freq > max_freq:
+                max_freq = freq
             if name is None:
                 name = clk_sig
-            xdc.append(
-                "create_clock -name {name} -period " + str(period) +
-                " [get_ports {clk}]".format(name=name, clk=clk_sig))
+            xdc.append("create_clock -name {name} -period {period} [get_ports {clk}]".format(
+                name   = name,
+                period = period,
+                clk    = clk_sig))
 
+        # FIXME: NextPNRWrapper is constructed at finalize level, too early
+        # to update self._pnr_opts. The solution is to update _nextpnr instance.
+        if max_freq > 0:
+            self._nextpnr._pnr_opts += f" --freq {round(max_freq, 3)}"
         # generate sdc
         xdc += self.additional_xdc_commands
         self._clock_constraints = "\n".join(xdc)
@@ -107,8 +122,7 @@ class XilinxYosysNextpnrToolchain(YosysNextPNRToolchain):
         if self.is_openxc7:
             chipdb_dir = os.environ.get('CHIPDB')
             if chipdb_dir is None or chipdb_dir == "":
-                print("Error: please specify the directory, where you store your nextpnr-xilinx chipdb files in the environment variable CHIPDB (directory may be empty)")
-                exit(1)
+                raise OSError("Error: please specify the directory, where you store your nextpnr-xilinx chipdb files in the environment variable CHIPDB (directory may be empty)")
         else:
             chipdb_dir = "/usr/share/nextpnr/xilinx-chipdb"
 
@@ -128,12 +142,13 @@ class XilinxYosysNextpnrToolchain(YosysNextPNRToolchain):
                 bba = self.dbpart + ".bba"
                 bbaexport = [pypy3, os.path.join(nextpnr_xilinx_python_dir, "bbaexport.py"), "--device", self.device, "--bba", bba]
                 print(str(bbaexport))
-                subprocess.run(bbaexport)
-                subprocess.run(["bbasm", "-l", bba, chipdb])
+                if subprocess.run(bbaexport).returncode != 0:
+                    raise OSError(f"Error occured during bbaexport's execution for '{chipdb}'.")
+                if subprocess.run(["bbasm", "-l", bba, chipdb]).returncode != 0:
+                    raise OSError(f"Error occured during bbasm's execution for '{chipdb}'.")
                 os.remove(bba)
             else:
-                print(f"Chip database file '{chipdb}' not found. Please check your toolchain installation!")
-                exit(1)
+                raise OSError(f"Chip database file '{chipdb}' not found. Please check your toolchain installation!")
 
         # pnr options
         self._pnr_opts += "--chipdb {chipdb} --write {top}_routed.json".format(
@@ -149,10 +164,9 @@ class XilinxYosysNextpnrToolchain(YosysNextPNRToolchain):
             prjxray_db_dir = "/usr/share/nextpnr/prjxray-db/"
 
         if not os.path.isdir(prjxray_db_dir):
-            print(f"{prjxray_db_dir} does not exist on your system. \n" + \
+            raise OSError(f"{prjxray_db_dir} does not exist on your system. \n" + \
                     "Do you have the openXC7 toolchain installed? \n" + \
                     "You can get it here: https://github.com/openXC7/toolchain-installer")
-            exit(1)
 
         # pre packer options
         self._pre_packer_opts[self._pre_packer_cmd[0]] = "--part {part} --db-root {db_root} {top}.fasm > {top}.frames".format(
@@ -169,6 +183,16 @@ class XilinxYosysNextpnrToolchain(YosysNextPNRToolchain):
 
         return YosysNextPNRToolchain.finalize(self)
 
+    def build_script(self):
+        build_filename = YosysNextPNRToolchain.build_script(self)
+        # Zynq7000/ZynqMP specific (signed bitstream).
+        if self.platform.device[0:4] in ["xc7z", "xczu"]:
+            with open(build_filename, "a") as fd:
+                script_contents = signed_bitstream_script(self.platform, self._build_name)
+                fd.write(script_contents)
+
+        return build_filename
+
     def build(self, platform, fragment,
         enable_xpm = False,
         **kwargs):
@@ -180,4 +204,14 @@ class XilinxYosysNextpnrToolchain(YosysNextPNRToolchain):
 
     def add_false_path_constraint(self, platform, from_, to):
         # FIXME: false path constraints are currently not supported by the toolchain
+        print("WARNING: false path constraints are not supported by the yosys+nextpnr toolchain and are ignored.")
         return
+
+
+def xilinx_yosys_nextpnr_args(parser):
+    toolchain_group = parser.add_argument_group(title="Yosys/NextPNR toolchain options")
+    yosys_nextpnr_args(toolchain_group)
+
+
+def xilinx_yosys_nextpnr_argdict(args):
+    return yosys_nextpnr_argdict(args)
